Time: 0 ms (100%), Space: 23.7 MB (22.16%) - LeetHub

+++ 0155-min-stack/0155-min-stack.cpp
@@ -16,7 +16,7 @@ public:
                 st.push(val);
             }
             else{               // val<minval
-                st.push(2*val - minval);   //minval = prevminval
+                st.push((long long)2*val - minval);   //minval = prevminval
                 minval=val;
             }
         }
